Use MariaDB, Python3 and latest Django.
All versioning has been removed. The benefit is that you will always be
running the latest and greatest of everything from the start. Because
the container django:latest is pretty much guaranteed to work the risk
that this will ever not work is very small.

Django extensions and ipython have been added to make development
easier.